microservices/payment/pyproject.toml
[tool.poetry]
name = "payment"
version = "0.1.0"
description = ""
authors = ["Clariteia Devs <devs@clariteia.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
minos-microservice-common = "^0.3.2"
minos-microservice-networks = { git = "https://github.com/Clariteia/minos_microservice_networks.git", branch = "issue-466-in-memory-request" }
minos-microservice-aggregate = "^0.2.2"
minos-microservice-saga = "^0.3.5"
minos-microservice-cqrs = "^0.2.1"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
black = "^19.10b"
isort = "^5.8.0"
pytest = "^6.2.4"
flake8 = "^3.9.2"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
microservice = "src.cli:main"